updating service

+++ src/domain/enrolls/enroll_dto.go
@@ -1,12 +1,12 @@
 package enrolls
 
 type Course struct {
-	CourseID int64        `json:"course_id"`
+	CourseID int          `json:"course_id"`
 	Users    []CourseUser `json:"users"`
 }
 
 type CourseUser struct {
-	UserID    int64  `json:"user_id"`
+	UserID    int    `json:"user_id"`
 	Username  string `json:"username"`
 	Firstname string `json:"firstname"`
 	Surname   string `json:"surname"`
@@ -14,8 +14,8 @@ type CourseUser struct {
 }
 
 type Enroll struct {
-	ID       int64  `json:"id"`
-	UserID   int64  `json:"user_id"`
-	CourseID int64  `json:"course_id"`
+	ID       int    `json:"id"`
+	UserID   int    `json:"user_id"`
+	CourseID int    `json:"course_id"`
 	Cohort   string `json:"cohort"`
 }
